\name{RBFNN}
\alias{RBFNN}
\docType{package}
\title{
  Radial Basis Function Neural Networks
}
\description{
  Package for using RBFNN. It can handle the original dataset, distance matrix or Gaussian kernel matrix as input types.
}
\details{
  TO DO: This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Xavier Martín Ballesteros, xmartin46@gmail.com

Maintainer: Xavier Martín Ballesteros <xmartin46@gmail.com>
}
\references{
  TO DO: References
}
\keyword{ RBFNN }
\seealso{
  TO DO: Link with train and predict.
}
\examples{
  \dontrun{
     TO DO
  }
}
\encoding{UTF-8}
