% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBFNN.R
\name{RBFNN_train}
\alias{RBFNN_train}
\title{Train a Radial Basis Function Neural Network.}
\usage{
RBFNN_train(X, Y, neurons, it, inputType = "original_dataset")
}
\arguments{
\item{X}{original dataset, distance matrix or gaussian kernel matrix.}

\item{Y}{the labels of \code{X}. Each observation has one label.}

\item{neurons}{number of neurons in the hidden layer.}

\item{it}{value used in the seed.}

\item{inputType}{type of \code{X}. Available types: original_dataset (original dataset), distance_matrix (Euclidean distance matrix ) and kernel_matrix (Gaussian kernel matrix).}
}
\value{
The trained model, the centroids of the RBFNN, the standard deviations of each RBF neuron and indexes of the observations that are the centroids.
}
\description{
Train a Radial Basis Function Neural Network.
}
\examples{
data(iris)
dataset <- iris[, 1:4]
labels <- iris[, 5]

X_train <- dataset[1:100, ]
y_train <- labels[1:100]
X_test <- dataset[101:150, ]
y_test <- labels[101:150]

neurons <- 10
seed_val <- 1
RBFNN <- RBFNN_train(as.matrix(X_train), as.matrix(y_train), neurons, seed_val)
}
